feat(rules): 为规则系统添加唯一ID标识并优化回测实体结构
- 在AbstractRule中添加基于SnowFlake的唯一ID生成
- 为Rule接口添加getId()方法定义
- 更新BacktestIndicator实体的指标值类型字段为整型枚举
- 将BacktestPositions中的持仓量字段从Long改为Integer
- 修改BacktestRule中satisfied字段从整型改为布尔类型
- 重构BacktestTradeLog中的成交量字段类型
- 在BaseTest中添加完整的回测记录创建和保存功能
- 创建Constant接口定义指标值类型的常量枚举

+++ clever-task-example/src/test/java/quant/entity/BacktestPositions.java
@@ -32,11 +32,11 @@ public class BacktestPositions {
     /**
      * 持有总量
      */
-    private Long volume;
+    private Integer volume;
     /**
      * 可用量
      */
-    private Long availableVolume;
+    private Integer availableVolume;
     /**
      * 平均成本价
      */
